refactor: extract PreviewComponent into a separate composable
