+++ design_system/src/commonMain/kotlin/net/thechance/mena/designsystem/presentation/component/datePicker/WheelDatePicker.kt
@@ -223,7 +223,8 @@ private fun VerticalPicker(
     }
 }
 
-@Preview
+@Preview(showBackground = true,backgroundColor = 0xFFF2F4F7)
+
 @Composable
 private fun WheelDatePickerPreview() {
     MenaTheme {
