+++ design_system/src/commonMain/kotlin/net/thechance/mena/designsystem/presentation/component/dialog/Dialog.kt
@@ -156,7 +156,7 @@ fun ScaffoldScope.Dialog(
     }
 }
 
-@Preview
+@Preview(showBackground = true,backgroundColor = 0xFFF2F4F7)
 @Composable
 private fun DialoguePreview() {
     MenaTheme {
