+++ design_system/src/commonMain/kotlin/net/thechance/mena/designsystem/presentation/component/indicator/DotsProgressIndicator.kt
@@ -75,7 +75,7 @@ fun DotsProgressIndicator(
 }
 
 
-@Preview
+@Preview(showBackground = true,backgroundColor = 0xFFF2F4F7)
 @Composable
 private fun DotsProgressIndicatorPreview() {
     Box(
