+++ design_system/src/commonMain/kotlin/net/thechance/mena/designsystem/presentation/component/progressBar/PorgressBar.kt
@@ -53,7 +53,7 @@ fun ProgressBar(
     )
 }
 
-@Preview
+@Preview(showBackground = true,backgroundColor = 0xFFF2F4F7)
 @Composable
 private fun ProgressBarPreview() {
     MenaTheme {
